+++ packages/sdk-core/src/interfaces.ts
@@ -47,9 +47,16 @@ export interface ISuperTokenModifyFlowParams {
 export interface ISuperTokenCreateFlowParams
     extends ISuperTokenModifyFlowParams {
     readonly flowRate: string;
+}
+
+export interface ISuperTokenCreateFlowByOperatorParams
+    extends ISuperTokenCreateFlowParams {
     readonly sender: string;
 }
+
 export type ISuperTokenUpdateFlowParams = ISuperTokenCreateFlowParams;
+export type ISuperTokenUpdateFlowByOperatorParams =
+    ISuperTokenCreateFlowByOperatorParams;
 export interface ISuperTokenDeleteFlowParams
     extends ISuperTokenModifyFlowParams {
     readonly sender: string;
@@ -107,21 +114,47 @@ export interface ISuperTokenUpdateSubscriptionUnitsParams {
     readonly userData?: string;
     readonly overrides?: Overrides & { from?: string | Promise<string> };
 }
-export interface IModifyFlowParams {
-    readonly receiver: string;
+
+export interface IModifyFlowParams extends ISuperTokenModifyFlowParams {
     readonly superToken: string;
-    readonly flowRate?: string;
-    readonly sender?: string;
-    readonly userData?: string;
-    readonly overrides?: Overrides & { from?: string | Promise<string> };
 }
 export interface ICreateFlowParams extends IModifyFlowParams {
     readonly flowRate: string;
 }
+export interface ICreateFlowByOperatorParams extends ICreateFlowParams {
+    readonly sender: string;
+}
 
 export type IUpdateFlowParams = ICreateFlowParams;
+export type IUpdateFlowByOperatorParams = ICreateFlowByOperatorParams;
+
 export interface IDeleteFlowParams extends IModifyFlowParams {
     readonly sender: string;
+}
+
+export interface ISuperTokenUpdateFlowOperatorPermissionsParams {
+    readonly sender: string;
+    readonly flowOperator: string;
+    readonly permissions: number;
+    readonly flowRateAllowance: string;
+    readonly userData?: string;
+    readonly overrides?: Overrides & { from?: string | Promise<string> };
+}
+
+export interface ISuperTokenFullControlParams {
+    readonly sender: string;
+    readonly flowOperator: string;
+    readonly userData?: string;
+    readonly overrides?: Overrides & { from?: string | Promise<string> };
+}
+
+export interface IUpdateFlowOperatorPermissionsParams
+    extends ISuperTokenUpdateFlowOperatorPermissionsParams {
+    readonly superToken: string;
+}
+
+export interface IFullControlParams extends ISuperTokenFullControlParams {
+    readonly superToken: string;
 }
 
 export interface IRealtimeBalanceOfParams {
@@ -164,6 +197,29 @@ export interface IGetFlowParams {
 export interface IGetAccountFlowInfoParams {
     readonly superToken: string;
     readonly account: string;
+    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+}
+
+export interface IGetFlowOperatorDataParams {
+    readonly superToken: string;
+    readonly sender: string;
+    readonly flowOperator: string;
+    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+}
+
+export interface IGetFlowOperatorDataByIDParams {
+    readonly superToken: string;
+    readonly flowOperatorId: string;
+    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+}
+export interface ISuperTokenFlowOperatorDataParams {
+    readonly sender: string;
+    readonly flowOperator: string;
+    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+}
+
+export interface ISuperTokenFlowOperatorDataByIDParams {
+    readonly flowOperatorId: string;
     readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
 }
 
@@ -415,4 +471,15 @@ export interface IWeb3FlowInfo {
     readonly flowRate: string;
     readonly deposit: string;
     readonly owedDeposit: string;
+}
+
+export interface IWeb3FlowOperatorDataParams {
+    readonly flowOperatorId: string;
+    readonly permissions: number;
+    readonly flowRateAllowance: ethers.BigNumber;
+}
+export interface IWeb3FlowOperatorData {
+    readonly flowOperatorId: string;
+    readonly permissions: string;
+    readonly flowRateAllowance: string;
 }