+++ packages/subgraph/test/queries/holQueries.ts
@@ -43,6 +43,7 @@ export const getStream = gql`
             id
             currentFlowRate
             streamedUntilUpdatedAt
+            deposit
             token {
                 id
             }
@@ -93,6 +94,29 @@ export const getStreamPeriod = gql`
             }
 
             totalAmountStreamed
+        }
+    }
+`;
+
+export const getFlowOperator = gql`
+    query getFlowOperator($id: ID!) {
+        response: flowOperator(id: $id) {
+            id
+            permissions
+            flowRateAllowanceGranted
+            flowRateAllowanceRemaining
+            sender {
+                id
+            }
+            token {
+                id
+            }
+            accountTokenSnapshot {
+                id
+            }
+            flowOperatorUpdatedEvents(orderBy: timestamp, orderDirection: asc) {
+                id
+            }
         }
     }
 `;