+++ packages/subgraph/test/validation/hol/streamValidator.ts
@@ -1,19 +1,14 @@
 import { BigNumber } from "@ethersproject/bignumber";
 import { expect } from "chai";
 import { fetchEntityAndEnsureExistence } from "../../helpers/helpers";
-import {
-    IStream,
-    IStreamData,
-    IEvent,
-    IAccount,
-} from "../../interfaces";
+import { IStream, IStreamData, IEvent, IAccount } from "../../interfaces";
 import { getAccount, getStream } from "../../queries/holQueries";
 import { validateReverseLookup } from "../validators";
 
 export const fetchStreamAndValidate = async (
     streamData: IStreamData,
     expectedStreamedUntilUpdatedAt: BigNumber,
-    flowRate: string,
+    newFlowRate: string,
     event: IEvent,
     isCreate: boolean
 ) => {
@@ -28,7 +23,7 @@ export const fetchStreamAndValidate = async (
         stream,
         expectedStreamedUntilUpdatedAt.toString(),
         streamId,
-        flowRate
+        newFlowRate
     );
 
     // validate flowUpdated reverse lookup on Stream entity
@@ -36,11 +31,11 @@ export const fetchStreamAndValidate = async (
 
     if (isCreate) {
         // validate accounts reverse lookup on Stream entity creation
-        validateAccountReverseLookups(stream);
+        await validateAccountReverseLookupsForStream(stream);
     }
 };
 
-export const validateAccountReverseLookups = async (stream: IStream) => {
+const validateAccountReverseLookupsForStream = async (stream: IStream) => {
     const senderAccount = await fetchEntityAndEnsureExistence<IAccount>(
         getAccount,
         stream.sender.id,
@@ -56,19 +51,19 @@ export const validateAccountReverseLookups = async (stream: IStream) => {
     validateReverseLookup(streamLightEntity, receiverAccount.inflows);
 };
 
-export const validateStreamEntity = (
+const validateStreamEntity = (
     subgraphStream: IStream,
     expectedStreamedUntilUpdatedAt: string,
     streamId: string,
-    currentFlowRate: string
+    newFlowRate: string
 ) => {
-    expect(subgraphStream.id, "Stream: id error").to.be.equal(streamId);
+    expect(subgraphStream.id, "Stream: id error").to.equal(streamId);
     expect(
         subgraphStream.currentFlowRate,
         "Stream: currentFlowRate error"
-    ).to.equal(currentFlowRate);
+    ).to.equal(newFlowRate);
     expect(
         subgraphStream.streamedUntilUpdatedAt,
         "Stream: streamedUntilUpdatedAt error"
-    ).to.be.equal(expectedStreamedUntilUpdatedAt);
+    ).to.equal(expectedStreamedUntilUpdatedAt);
 };