+++ packages/subgraph/test/validation/hol/tokenValidator.ts
@@ -35,22 +35,20 @@ export const validateTokenEntity = (
     expectedUnderlyingAddress: string,
     expectedDecimals: number
 ) => {
-    expect(subgraphToken.name, "SuperToken: name error").to.be.equal(
-        expectedName
-    );
+    expect(subgraphToken.name, "SuperToken: name error").to.equal(expectedName);
     expect(subgraphToken.symbol, "SuperToken: symbol error").to.equal(
         expectedSymbol
     );
-    expect(subgraphToken.isListed, "SuperToken: isListed error").to.be.equal(
+    expect(subgraphToken.isListed, "SuperToken: isListed error").to.equal(
         expectedIsListed
     );
-    expect(subgraphToken.decimals).to.be.equal(expectedDecimals);
+    expect(subgraphToken.decimals).to.equal(expectedDecimals);
 
     if (subgraphToken.underlyingAddress === "0x") {
-        expect(subgraphToken.underlyingToken).to.be.equal(null);
+        expect(subgraphToken.underlyingToken).to.equal(null);
     } else {
         // NOTE: underlyingToken should not be null here
-        expect(subgraphToken.underlyingToken!.id).to.be.equal(
+        expect(subgraphToken.underlyingToken!.id).to.equal(
             expectedUnderlyingAddress.toLowerCase()
         );
     }